// wren-ui/src/types/api.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DialectSQL: any;
  JSON: any;
};
// ...
export type CreateInstructionInput = {
  instruction: Scalars['String'];
  isDefault: Scalars['Boolean'];
  questions: Array<Scalars['String']>;
};
// ...
export type CreateSqlPairInput = {
  question: Scalars['String'];
  sql: Scalars['String'];
};
// ...
export type Instruction = {
  __typename?: 'Instruction';
  createdAt: Scalars['String'];
  id: Scalars['Int'];
  instruction: Scalars['String'];
  isDefault: Scalars['Boolean'];
  questions: Array<Scalars['String']>;
  updatedAt: Scalars['String'];
};
// ...
export type SqlPair = {
  __typename?: 'SqlPair';
  createdAt?: Maybe<Scalars['String']>;
  id: Scalars['Int'];
  question: Scalars['String'];
  sql: Scalars['String'];
  updatedAt?: Maybe<Scalars['String']>;
};